🎨 organize components file
